refactor: error dir
