+++ src/app/v1/users/handler/users_handler.rs
@@ -3,7 +3,7 @@ use std::ops::Deref;
 use actix_web::{post, web, HttpResponse, Responder};
 use slog::{crit, o};
 
-use crate::app::error::error::AppError;
+use crate::app::error::AppError;
 use crate::app::v1::users::dto::post_users_request::PostUsersRequest;
 use crate::app::v1::users::service::users_service;
 use crate::app::AppState;
