+++ src/app/v1/users/service/users_service.rs
@@ -1,4 +1,4 @@
-use crate::app::error::error::AppError;
+use crate::app::error::AppError;
 use crate::app::v1::users::dto::post_users_request::PostUsersRequest;
 use crate::app::v1::users::model::user::User;
 use crate::app::v1::users::repository::user_repository;
